Move from Apache HTTPClient5 alpha1 to alpha2 to fix the remove-headers problem.
